app/src/test/java/com/kristoffersol/materialtimer/InjectorUtilsTest.java
package com.kristoffersol.materialtimer;

import com.kristoffersol.materialtimer.data.PomodoroRepository;
import com.kristoffersol.materialtimer.util.InjectorUtils;
import com.kristoffersol.materialtimer.viewmodel.PomodoroViewModelFactory;

import org.junit.Before;
import org.junit.Rule;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.Mock;
import org.mockito.Mockito;
import org.mockito.MockitoAnnotations;
import org.powermock.api.mockito.PowerMockito;
import org.powermock.core.classloader.annotations.PrepareForTest;
import org.powermock.modules.junit4.PowerMockRunner;
import org.powermock.modules.junit4.rule.PowerMockRule;

import static junit.framework.TestCase.assertEquals;

@RunWith( PowerMockRunner.class )
@PrepareForTest( InjectorUtils.class )
public class InjectorUtilsTest {

//    @Rule
//    public PowerMockRule rule = new PowerMockRule();

    @Mock
    private PomodoroViewModelFactory factory;

    @Mock
    private PomodoroRepository repository;

    @Before
    public void setUp(){
        MockitoAnnotations.initMocks(this);
//        PowerMockito.mockStatic(InjectorUtils.class);
    }

    @Test
    public void shouldReturnViewModelFactory(){
//        // set expectations
//        Mockito.when(InjectorUtils.providePomodoroViewModelFactory()).thenReturn(factory);
//
//        // verify behavior
//        PowerMockito.verifyStatic(InjectorUtils.class);
//        InjectorUtils.providePomodoroViewModelFactory();
        assertEquals(1,1);
    }

    @Test
    public void shouldReturnPomodoroRepository(){
//
//        Mockito.when(InjectorUtils.providePomodoroRepository()).thenReturn(repository);
//
//        PowerMockito.verifyStatic(InjectorUtils.class);
//        InjectorUtils.providePomodoroRepository();

        assertEquals(1,1);
    }
}
